backend/requirements.txt
fastapi==0.115.0
uvicorn>=0.34.0
pandas==2.2.3
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic>=2.11.2
anyio>=4.8.0
httpx>=0.28.1
websockets>=13.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google ADK and AI dependencies
google-genai
google-adk
streamlit
plotly
